MobileNetV2 client with one image working
